@@ -1,21 +1,15 @@
 cmake_minimum_required(VERSION 3.1)
 project(embedded VERSION 0.1.0)
 
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++20")
 set(CMAKE_CXX_EXTENSIONS OFF)
 set(Boost_USE_STATIC_LIBS ON)
 set(BITCOIN_FOLDER "contrib/bitcoin")
-if(ANDROID)
-    if (${ANDROID_ABI} STREQUAL "armeabi-v7a")
-        set(BITCOIN_FOLDER "contrib/bitcoin-arm8")
-    endif ()
-elseif(NOT ANDROID)
-    find_package(Threads REQUIRED)
-    find_package(Boost 1.47.0 REQUIRED COMPONENTS filesystem)
-endif()
-
-set(UR__DISABLE_TESTS ON)
+   
+find_package(Threads REQUIRED)
+find_package(Boost 1.47.0 REQUIRED)
 
 if(WIN32)
     find_package(Libevent CONFIG REQUIRED)
