@@ -1,8 +1,9 @@
 cmake_minimum_required(VERSION 3.1)
 project(nunchuk-examples VERSION 0.1.0)
 
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++20")
 set(CMAKE_CXX_EXTENSIONS OFF)
 
 add_subdirectory(.. lib EXCLUDE_FROM_ALL)
