@@ -20,25 +20,24 @@
 #include <chainparams.h>
 #include <pubkey.h>
 #include <validation.h>
+#include <kernel/chainparams.h>  // IWYU pragma: export
+#include <common/url.h>
 
 // required for util/translation.h
 const std::function<std::string(const char *)> G_TRANSLATION_FUN = nullptr;
 
-// required for validation.h
-static const ECCVerifyHandle verify_handle;
-
 EmbeddedRpc::EmbeddedRpc() {}
 
-EmbeddedRpc::~EmbeddedRpc() { ECC_Stop(); }
+EmbeddedRpc::~EmbeddedRpc() {}
 
 void EmbeddedRpc::Init(const std::string &chain) {
   static std::once_flag flag;
+  static ECC_Context ecc_context{};
   std::call_once(flag, [&] {
-    ECC_Start();
+    ECC_InitSanityCheck();
     chain_ = chain;
-    SelectParams(chain);
-    RegisterMiscRPCCommands(table_);
-    RegisterRawTransactionRPCCommands(table_);
+    SelectParams(ChainTypeFromString(chain).value());
+    RegisterAllCoreRPCCommands(table_);
     SetRPCWarmupFinished();
     initialized_ = true;
   });
@@ -48,7 +47,7 @@ void EmbeddedRpc::SetChain(const std::string &chain) {
   if (!initialized_) throw std::runtime_error("uninitialized");
   if (chain_ == chain) return;
   chain_ = chain;
-  SelectParams(chain_);
+  SelectParams(ChainTypeFromString(chain).value());
 }
 
 const std::string &EmbeddedRpc::GetChain() const {
@@ -64,9 +63,9 @@ std::string EmbeddedRpc::SendRequest(const std::string &body) const {
   req.parse(val_request);
   try {
     auto resp = table_.execute(req);
-    return JSONRPCReply(resp, NullUniValue, req.id);
+    return JSONRPCReplyObj(std::move(resp), NullUniValue, req.id, JSONRPCVersion::V2).write();
   } catch (const UniValue &err) {
-    return JSONRPCReply(NullUniValue, err, req.id);
+    return JSONRPCReplyObj(NullUniValue, err, req.id, JSONRPCVersion::V2).write();
   }
 }
 
