@@ -25,8 +25,8 @@
 #include <tap_protocol/utils.h>
 #include <utils/bip32.hpp>
 #include <utils/txutils.hpp>
+#include <spender.h>
 #include <utils/chain.hpp>
-#include "coinselector.h"
 #include "key_io.h"
 #include "nunchuk.h"
 #include "nunchukimpl.h"
@@ -825,7 +825,7 @@ static SatscardStatus GetSatscardstatus(tap_protocol::Satscard* satscard) {
       satscard->GetNumSlots(),      std::move(slots)};
 }
 
-static std::string GetSatscardSlotsDescriptor(
+static std::vector<std::string> GetSatscardSlotsDescriptor(
     const std::vector<SatscardSlot>& slots, bool use_privkey) {
   const auto get_slot_desc = [&](const SatscardSlot& slot) {
     if (use_privkey) {
@@ -842,50 +842,31 @@ static std::string GetSatscardSlotsDescriptor(
       }
 
       const std::string wif = EncodeSecret(key);
-      const std::string desc_wif = AddChecksum("wpkh(" + wif + ")");
-      return json({
-          {"desc", desc_wif},
-          {"internal", false},
-          {"active", true},
-      });
+      return AddChecksum("wpkh(" + wif + ")");
     }
-    if (slot.get_status() == SatscardSlot::Status::SEALED) {
-      if (slot.get_pubkey().empty()) {
-        throw NunchukException(NunchukException::INVALID_PARAMETER,
-                               "Invalid slot pubkey");
-      }
-
+    if (!slot.get_pubkey().empty()) {
       CPubKey pub(MakeUCharSpan(slot.get_pubkey()));
       if (!pub.IsValid()) {
         throw NunchukException(NunchukException::INVALID_PARAMETER,
                                "Invalid slot key");
       }
 
-      return json({
-          {"desc", AddChecksum("wpkh(" + HexStr(slot.get_pubkey()) + ")")},
-          {"internal", false},
-          {"active", true},
-      });
+      return AddChecksum("wpkh(" + HexStr(slot.get_pubkey()) + ")");
     }
-    return json({
-        {"desc", AddChecksum("addr(" + slot.get_address() + ")")},
-        {"internal", false},
-        {"active", true},
-    });
+    return AddChecksum("addr(" + slot.get_address() + ")");
   };
 
-  std::string desc = std::accumulate(std::begin(slots), std::end(slots), json(),
-                                     [&](json desc, const SatscardSlot& slot) {
-                                       desc.push_back(get_slot_desc(slot));
-                                       return desc;
-                                     })
-                         .dump();
-  return desc;
+  std::vector<std::string> ret(slots.size());
+  std::transform(slots.begin(), slots.end(), ret.begin(), get_slot_desc);
+
+  return ret;
 }
 
-static std::pair<Transaction, std::string> CreateSatscardSlotsTransaction(
-    const std::vector<SatscardSlot>& slots, const std::string& address,
-    const Amount& fee_rate, const Amount& discard_rate, bool use_privkey) {
+static std::pair<Transaction, std::vector<std::string>>
+CreateSatscardSlotsTransaction(const std::vector<SatscardSlot>& slots,
+                               const std::string& address,
+                               const Amount& fee_rate,
+                               const Amount& discard_rate, bool use_privkey) {
   std::vector<UnspentOutput> utxos;
   std::string change_address;
   Amount total_balance = 0;
@@ -907,54 +888,50 @@ static std::pair<Transaction, std::string> CreateSatscardSlotsTransaction(
     total_balance += slot.get_balance();
   }
 
-  std::vector<TxInput> selector_inputs;
   std::vector<TxOutput> selector_outputs{TxOutput{address, total_balance}};
+  std::vector<std::string> descs =
+      nunchuk::GetSatscardSlotsDescriptor(slots, use_privkey);
 
-  std::string desc = nunchuk::GetSatscardSlotsDescriptor(slots, use_privkey);
-
-  CoinSelector selector = [&]() -> CoinSelector {
-    if (use_privkey ||
-        (slots.size() == 1 &&
-         slots.front().get_status() == SatscardSlot::Status::SEALED)) {
-      auto ret = CoinSelector{desc, change_address};
-      ret.set_fee_rate(CFeeRate(fee_rate));
-      ret.set_discard_rate(CFeeRate(discard_rate));
-      return ret;
-    }
-    // No private key or pubkey for unsealed slot without cvc, only address
-    // so we use dummy script witness to estimate correct fee
-    CScriptWitness dummy_scriptwitness{};
-    dummy_scriptwitness.stack = {std::vector<unsigned char>(72),
-                                 std::vector<unsigned char>(33)};
-    auto ret = CoinSelector{CFeeRate(fee_rate), CFeeRate(discard_rate),
-                            std::move(dummy_scriptwitness)};
-    return ret;
-  }();
-
-  Amount fee = 0;
   int vsize = 0;
-  std::string error;
   int change_pos = 0;
-  if (!selector.Select(utxos, utxos, change_address, true, selector_outputs,
-                       selector_inputs, fee, vsize, error, change_pos)) {
+  auto res =
+      wallet::CreateTransaction(utxos, utxos, selector_outputs, true, descs,
+                                change_address, fee_rate, change_pos, vsize);
+  if (!res) {
+    std::string error = util::ErrorString(res).original;
     throw NunchukException(NunchukException::COIN_SELECTION_ERROR, error);
+  }
+  const auto& txr = *res;
+  CTransactionRef tx_new = txr.tx;
+
+  std::vector<TxInput> new_inputs;
+  for (const CTxIn& txin : tx_new->vin) {
+    new_inputs.push_back({txin.prevout.hash.GetHex(), txin.prevout.n});
+  }
+  std::vector<TxOutput> new_outputs;
+  for (const CTxOut& txout : tx_new->vout) {
+    CTxDestination address;
+    ExtractDestination(txout.scriptPubKey, address);
+    new_outputs.push_back({EncodeDestination(address), txout.nValue});
   }
 
   std::string base64_psbt =
-      CoreUtils::getInstance().CreatePsbt(selector_inputs, selector_outputs);
+      CoreUtils::getInstance().CreatePsbt(new_inputs, new_outputs);
 
+  Wallet wallet{false};
+  wallet.set_m(1);
   auto tx = GetTransactionFromPartiallySignedTransaction(
-      DecodePsbt(base64_psbt), {}, 1);
+      DecodePsbt(base64_psbt), wallet);
 
-  tx.set_fee(fee);
+  tx.set_fee(res->fee);
   tx.set_change_index(change_pos);
   tx.set_receive(false);
-  tx.set_sub_amount(total_balance - fee);
+  tx.set_sub_amount(total_balance - res->fee);
   tx.set_fee_rate(fee_rate);
   tx.set_subtract_fee_from_amount(true);
   tx.set_psbt(base64_psbt);
   tx.set_vsize(vsize);
-  return {std::move(tx), std::move(desc)};
+  return {std::move(tx), std::move(descs)};
 }
 
 SatscardStatus NunchukImpl::GetSatscardStatus(
@@ -1052,11 +1029,16 @@ Transaction NunchukImpl::SweepSatscardSlots(
     Amount fee_rate) {
   if (fee_rate <= 0) fee_rate = EstimateFee();
   auto discard_rate = synchronizer_->RelayFee();
-  auto [tx, desc] = nunchuk::CreateSatscardSlotsTransaction(
+  auto [tx, descs] = nunchuk::CreateSatscardSlotsTransaction(
       slots, address, fee_rate, discard_rate, true);
 
   auto psbt = DecodePsbt(tx.get_psbt());
-  auto provider = SigningProviderCache::getInstance().GetProvider(desc);
+  auto json_descs = json::array();
+  for (auto&& desc : descs) {
+    json_descs.push_back(desc);
+  }
+  auto provider =
+      SigningProviderCache::getInstance().GetProvider(json_descs.dump());
   int nin = psbt.tx.value().vin.size();
 
   std::vector<std::string> tx_ids;
