@@ -16,7 +16,7 @@
  */
 
 #include <nunchuk.h>
-#include <amount.h>
+#include <consensus/amount.h>
 
 #include <doctest.h>
 
