@@ -25,7 +25,7 @@ TEST_CASE("testing encrypt and descrypt attachment") {
   std::string body = "testtesttest";
   std::string accessToken = "";
   if (accessToken.empty()) return;
-  auto event_file = EncryptAttachment(accessToken, body);
-  CHECK(!event_file.empty());
-  CHECK(DecryptAttachment(event_file) == body);
+  // auto event_file = EncryptAttachment(accessToken, body);
+  // CHECK(!event_file.empty());
+  // CHECK(DecryptAttachment(event_file) == body);
 }
